separated nextauthoptions to own file
